Corregi la declaracion de srbk_handler

+++ TP2/Kernel/include/sbrk.h
@@ -2,7 +2,7 @@
 #define SBRK_H
 #include <stdint.h>
 
-void sbrk_handler(intptr_t increment, void ** buffer);
+void sbrk_handler(int increment, void ** buffer);
 int brk_handler( void * addr);
 
 #endif
